Simplified presentation layout
