+++ pyuavcan/presentation/__init__.py
@@ -6,5 +6,6 @@
 
 from ._presentation import Presentation
 
+from ._session import OutgoingTransferIDCounter
 from ._session import TypedSession, MessageTypedSession, ServiceTypedSession
 from ._session import Publisher, Subscriber
